// @types/contentful.d.ts
// THIS FILE IS AUTOMATICALLY GENERATED. DO NOT MODIFY IT.

import { Asset, Entry } from "contentful";
import { Document } from "@contentful/rich-text-types";

export interface IAlbumFields {
  /** title */
  title: string;

  /** slug */
  slug: string;

  /** songs */
  songs?: ISong[] | undefined;

  /** artist */
  artist?: IArtist | undefined;
}

export interface IAlbum extends Entry<IAlbumFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "album";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface IArtistFields {
  /** name */
  name: string;

  /** slug */
  slug: string;

  /** songs */
  songs?: ISong[] | undefined;

  /** albums */
  albums?: IAlbum[] | undefined;
}

export interface IArtist extends Entry<IArtistFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "artist";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface ISongFields {
  /** title */
  title: string;

  /** slug */
  slug: string;

  /** album */
  album?: IAlbum | undefined;

  /** artist */
  artist?: IArtist | undefined;
}

export interface ISong extends Entry<ISongFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "song";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export type CONTENT_TYPE = "album" | "artist" | "song";

export type LOCALE_CODE = "en-US";

export type CONTENTFUL_DEFAULT_LOCALE_CODE = "en-US";